refactor: Header -> AppHeader 변경
