+++ src/App.tsx
@@ -1,13 +1,13 @@
 import { BrowserRouter as Router } from "react-router-dom"
-import Header from "./widgets/ui/Header.tsx"
-import Footer from "./widgets/ui/Footer.tsx"
-import PostsManagerPage from "./pages/PostsManagerPage.tsx"
+import { AppHeader } from "@/widgets/app-header"
+import Footer from "@/widgets/ui/Footer.tsx"
+import PostsManagerPage from "@/pages/PostsManagerPage.tsx"
 
 const App = () => {
   return (
     <Router>
       <div className="flex flex-col min-h-screen">
-        <Header />
+        <AppHeader />
         <main className="flex-grow container mx-auto px-4 py-8">
           <PostsManagerPage />
         </main>
