+++ src/widgets/app-header/index.ts
@@ -1,0 +1,1 @@
+export { AppHeader } from "./ui"
