+++ src/widgets/app-header/ui/index.ts
@@ -1,0 +1,1 @@
+export { default as AppHeader } from "./AppHeader"
